+++ job-server-extras/src/main/scala/spark/jobserver/context/SessionContextFactory.scala
@@ -27,7 +27,6 @@ class SessionContextFactory extends ScalaContextFactory {
 
   def makeContext(sparkConf: SparkConf, config: Config, contextName: String): C = {
     val builder = SparkSession.builder()
-    builder.config("spark.sql.catalogImplementation", "hive")
     builder.config(sparkConf).appName(contextName)
     try {
       builder.enableHiveSupport()
