+++ job-server-extras/src/test/scala/spark/jobserver/SessionJobSpec.scala
@@ -17,7 +17,6 @@ class TestSessionContextFactory extends SessionContextFactory {
     builder.config(sparkConf).appName(contextName).master("local")
     builder.config("javax.jdo.option.ConnectionURL", "jdbc:derby:memory:myDB;create=true")
     builder.config("javax.jdo.option.ConnectionDriverName", "org.apache.derby.jdbc.EmbeddedDriver")
-    builder.config("spark.sql.catalogImplementation", "hive")
     try {
       builder.enableHiveSupport()
     } catch {
